Fix Windows tests

+++ archiver_test.go
@@ -25,7 +25,8 @@ func TestMkdirAll(t *testing.T) {
 	// Check that the directories were created with 0700 permissions
 	info, err := os.Stat(testPath)
 	assert.NoError(t, err, "Failed to stat directory")
-	assert.Equal(t, os.FileMode(0700), info.Mode().Perm(), "Expected directory to have permissions 0700")
+
+	assert.Equal(t, uint32(os.FileMode(0700)), uint32(info.Mode().Perm()), "Expected directory to have permissions 0700")
 
 	// Step 2: Run MkdirAll which should change permissions to 0755
 	restore, err := mkdirAll(testPath)
@@ -34,14 +35,14 @@ func TestMkdirAll(t *testing.T) {
 	// Check that the permissions were changed to 0755
 	info, err = os.Stat(testPath)
 	assert.NoError(t, err, "Failed to stat directory after MkdirAll")
-	assert.Equal(t, os.FileMode(0755), info.Mode().Perm(), "Expected directory to have permissions 0755 after MkdirAll")
+	assert.Equal(t, uint32(os.FileMode(0755)), uint32(info.Mode().Perm()), "Expected directory to have permissions 0755 after MkdirAll")
 
 	// Step 3: Simulate changing the permissions to 0700 (simulating a permission modification)
 	err = os.Chmod(testPath, 0700)
 	assert.NoError(t, err, "Failed to change directory permissions to 0700")
 	info, err = os.Stat(testPath)
 	assert.NoError(t, err, "Failed to stat directory after permission change")
-	assert.Equal(t, os.FileMode(0700), info.Mode().Perm(), "Expected directory to have permissions 0700 after permission change")
+	assert.Equal(t, uint32(os.FileMode(0700)), uint32(info.Mode().Perm()), "Expected directory to have permissions 0700 after permission change")
 
 	// Step 4: Restore the original permissions using the deferred restore function
 	err = restore()
@@ -50,7 +51,7 @@ func TestMkdirAll(t *testing.T) {
 	// Verify that the original permissions (0700) were restored
 	info, err = os.Stat(testPath)
 	assert.NoError(t, err, "Failed to stat directory after restoration")
-	assert.Equal(t, os.FileMode(0700), info.Mode().Perm(), "Expected directory to have restored permissions 0755")
+	assert.Equal(t, uint32(os.FileMode(0700)), uint32(info.Mode().Perm()), "Expected directory to have restored permissions 0755")
 }
 
 func TestWithin(t *testing.T) {
